Return error from _report_local_error
A call to `_report_local_error` is always followed immediately by a
`raise` of the newly recorded error class and message (this is not the
case of `_report_remote_error`, only `_report_local_error`). To simplify
this a bit, construct the exception object in `_report_local_error`,
return it, and allow the caller to raise that object rather than
constructing one in each place this is done.