Split the receive and transmit state machines
Allow interleaving of request and response data, so that it is no longer
necessary to read the full request body before starting the response.
The SCGI specification technically prohibits sending any response data
until the server has seen the whole request, but doing so is permitted
by CGI and works fine with some SCGI clients, therefore it makes sense
for an SCGI server to make its own choice whether or not to do that
based on the environment rather than enforcing a blanket prohibition.