Delete unnecessary check
The netstring length prefix is now checked for a more specific sane
length limit. In every other receive state machine state, this check is
unnecessary: in `HEADERS` we already know that we will stop and decode
the netstring after `_rx_env_length` (which is itself known to meet the
limit), and in `BODY` we dispose of all buffered data immediately.